token相关

+++ dlfc-zfw-wechat-house/src/main/java/com/dlfc/services/house/controller/HouLeaseController.java
@@ -80,14 +80,17 @@ public class HouLeaseController {
     @RequestMapping(method = RequestMethod.POST)
     public ListResultDTO<HouseDTO> houses(@RequestBody HouseConditionDTO conditionDTO,
                                           @RequestHeader(required = false) String token) throws CustomRuntimeException {
-        HouLeaseInfoDTO dto = null;
+        getUser(token);
+        if (null == user) {
+            ListResultDTO.failure(new ArrayList<>(), resultError);
+        }
+        HouLeaseInfoDTO dto;
         try {
             dto = conditionConvertor.toModel(conditionDTO);
             houLeaseInfoList = houseLeaseInfoService.findByParams(dto);
             if (null == houLeaseInfoList || houLeaseInfoList.size() == 0) {
                 return houseInfoConvertor.toResultDTO(new ArrayList<HouLeaseInfo>());
             }
-            getUser(token);
             return houseInfoConvertor.toResultDTO(houLeaseInfoList, user.getId());
         } catch (Exception e) {
             return houseInfoConvertor.toResultDTO(houLeaseInfoList);
@@ -104,6 +107,9 @@ public class HouLeaseController {
     public ResultDTO<Void> update(@RequestBody HouseDTO houseDTO,
                                   @RequestHeader String token) {
         getUser(token);
+        if (null == user) {
+            return ResultDTO.failure(resultError);
+        }
         SysTrafficLines sysTrafficLines;
         SysDescriptions sysDescriptions;
         SysInfoAtt sysInfoAtt;
@@ -190,6 +196,9 @@ public class HouLeaseController {
         List<HouLeaseInfo> houLeaseInfos = houseLeaseInfoService.findAll(pageNo, pageSize);
         try {
             getUser(token);
+            if (null == user) {
+                return ListResultDTO.failure(new ArrayList<HouseDTO>(), resultError);
+            }
         } catch (Exception e) {
             return houseInfoConvertor.toResultDTO(houLeaseInfos);
         }
@@ -242,6 +251,9 @@ public class HouLeaseController {
         try {
             houLeaseInfo = houseLeaseInfoService.findByHouseLeaseInfo(lid);
             getUser(token);
+            if (null == user) {
+                return ResultDTO.failure(new HouseDTO(), resultError);
+            }
             if (houLeaseInfo == null) {
                 return null;
             }
@@ -255,11 +267,14 @@ public class HouLeaseController {
      * 删除房源
      */
     @RequestMapping(value = "/delete", method = RequestMethod.DELETE)
-    public ResultDTO delete(@RequestParam String lid, @RequestHeader String token) {
+    public ResultDTO<Void> delete(@RequestParam String lid, @RequestHeader String token) {
         ResultError resultError = null;
         try {
             if (StringUtils.isNotEmpty(token)) {
                 getUser(token);
+                if (null == user) {
+                    return ResultDTO.failure(resultError);
+                }
                 boolean success = houseLeaseInfoService.delete(lid);
                 if (success) {
                     return ResultDTO.success();
